+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Canvas Calendar - arrows for changing week, marker for each day which has events, possibility for saving events [#1274](https://github.com/rokwire/illinois-app/issues/1274).
 ### Changed
 - Inbox renamed to Notifications [#1326](https://github.com/rokwire/illinois-app/issues/1326).
+- In Groups allow an Admin to End/Close a Poll even if they did not create it. [#1328](https://github.com/rokwire/illinois-app/issues/1328).
 ### Removed
 - Canvas Course code from the card [#1274](https://github.com/rokwire/illinois-app/issues/1274).